feat: remove FirebaseDatabaseProvider and FirebaseDtabaseMutation from admin/editor
